requirements.txt
xlwt==1.3.0
django==2.1.3
djangorestframework==3.9.0
